+++ src/editor/console/components/__tests__/console-pane.test.js
