+++ src/editor/console/components/base-icon.jsx
