+++ src/editor/console/components/console-input.jsx
