+++ src/editor/console/components/double-chevron-icon.js
